separate scalar.rs to single file
